moved device widgets and plots into their own files inside new devices and plots directories, changed nothing functionally
