+++ src/app.py
@@ -47,8 +47,27 @@ from widgets import (
     StepsWidget,
     FloatTextEdit,
     StatusLights,
-    PulseQuality
 )
+from plots import (
+    MainPlot,
+    SinglePlot, 
+    TriplePlot, 
+    AFMPlot, 
+    ElectrometerPlot
+)
+
+from devices import (
+    CPCWidget, 
+    PSMWidget, 
+    CO2Widget, 
+    ElectrometerWidget, 
+    RHTPWidget,
+    eDiluterWidget, 
+    AFMWidget, 
+    TSIWidget, 
+    ExampleDeviceWidget
+)
+
 from serial_connection import SerialDeviceConnection
 from params import ScalableGroup, params, p
 
@@ -2727,1153 +2746,6 @@ class MainWindow(QMainWindow):
         print("Restarting timer...")
         self.startTimer() # start timer
         self.timer_functions() # call timer functions at start time
-
-# main plot widget
-class MainPlot(GraphicsLayoutWidget):
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-        
-        # create plot by adding it to widget
-        self.plot = self.addPlot()
-        # create legend
-        self.legend = LegendItem(offset=(70,20), labelTextColor='w', labelTextSize='11pt')
-        self.legend.setParentItem(self.plot.graphicsItem())
-        # create dictionaries for viewboxes and axes, use device type as key
-        self.viewboxes = {}
-        self.axes = {}
-
-        # time axis
-        self.plot.setAxisItems({'bottom':DateAxisItem()}) # set time axis to bottom
-        self.plot.setLabel('bottom', "Time") # set time axis label
-        self.axis_time = self.plot.getAxis('bottom') # store time axis to variable
-        self.set_axis_style(self.axis_time, 'w') # set axis style
-        self.axis_time.enableAutoSIPrefix(enable=False) # disable auto SI prefix
-
-        # CPC viewbox
-        self.viewboxes[CPC] = self.plot.getViewBox() # store default viewbox to dictionary
-        # CPC axis
-        self.axes[CPC] = self.plot.getAxis('left') # store left axis to dictionary
-        self.axes[CPC].setLabel('CPC concentration', units='#/cc', color='w') # set label
-        self.set_axis_style(self.axes[CPC], 'w') # set axis style
-
-        # PSM viewbox # TODO create function for viewbox and axis creation
-        self.viewboxes[PSM] = ViewBox() # create viewbox
-        self.plot.scene().addItem(self.viewboxes[PSM]) # add viewbox to scene
-        self.viewboxes[PSM].setXLink(self.plot) # link x axis of viewbox to x axis of plot
-        # PSM axis
-        self.axes[PSM] = AxisItem('right') # create second axis
-        self.plot.layout.addItem(self.axes[PSM], 2, 3) # add axis to plot
-        self.axes[PSM].setLabel('PSM saturator flow rate', units='lpm', color='w') # set label
-        self.set_axis_style(self.axes[PSM], 'w') # set axis style
-        self.axes[PSM].linkToView(self.viewboxes[PSM]) # link axis to viewbox
-
-        # ELECTROMETER viewbox
-        self.viewboxes[ELECTROMETER] = ViewBox() # create viewbox
-        self.plot.scene().addItem(self.viewboxes[ELECTROMETER]) # add viewbox to scene
-        self.viewboxes[ELECTROMETER].setXLink(self.plot) # link x axis of viewbox to x axis of plot
-        # ELECTROMETER axis
-        self.axes[ELECTROMETER] = AxisItem('right') # create third axis
-        self.plot.layout.addItem(self.axes[ELECTROMETER], 2, 4) # add axis to plot
-        self.axes[ELECTROMETER].setLabel('ELECTROMETER voltage 2', units='V', color='w') # set label
-        self.set_axis_style(self.axes[ELECTROMETER], 'w') # set axis style
-        self.axes[ELECTROMETER].linkToView(self.viewboxes[ELECTROMETER]) # link axis to viewbox
-
-        # CO2 viewbox
-        self.viewboxes[CO2_SENSOR] = ViewBox() # create viewbox
-        self.plot.scene().addItem(self.viewboxes[CO2_SENSOR]) # add viewbox to scene
-        self.viewboxes[CO2_SENSOR].setXLink(self.plot) # link x axis of viewbox to x axis of plot
-        # CO2 axis
-        self.axes[CO2_SENSOR] = AxisItem('right') # create fourth axis
-        self.plot.layout.addItem(self.axes[CO2_SENSOR], 2, 5) # add axis to plot
-        self.axes[CO2_SENSOR].setLabel('CO2 concentration', units='ppm', color='w') # set label
-        self.set_axis_style(self.axes[CO2_SENSOR], 'w') # set axis style
-        self.axes[CO2_SENSOR].linkToView(self.viewboxes[CO2_SENSOR]) # link axis to viewbox
-
-        # RHTP viewbox
-        self.viewboxes[RHTP] = ViewBox() # create viewbox
-        self.plot.scene().addItem(self.viewboxes[RHTP]) # add viewbox to scene
-        self.viewboxes[RHTP].setXLink(self.plot) # link x axis of viewbox to x axis of plot
-        # RHTP axis
-        self.axes[RHTP] = AxisItem('right') # create fifth axis
-        self.plot.layout.addItem(self.axes[RHTP], 2, 6) # add axis to plot
-        self.axes[RHTP].setLabel('RHTP', color='w') # set label
-        self.set_axis_style(self.axes[RHTP], 'w') # set axis style
-        self.axes[RHTP].linkToView(self.viewboxes[RHTP]) # link axis to viewbox
-
-        # AFM viewbox
-        self.viewboxes[AFM] = ViewBox() # create viewbox
-        self.plot.scene().addItem(self.viewboxes[AFM]) # add viewbox to scene
-        self.viewboxes[AFM].setXLink(self.plot)
-        # AFM axis
-        self.axes[AFM] = AxisItem('right') # create sixth axis
-        self.plot.layout.addItem(self.axes[AFM], 2, 7) # add axis to plot
-        self.axes[AFM].setLabel('AFM', color='w')
-        self.set_axis_style(self.axes[AFM], 'w')
-        self.axes[AFM].linkToView(self.viewboxes[AFM])
-
-        # eDiluter viewbox
-        self.viewboxes[EDILUTER] = ViewBox() # create viewbox
-        self.plot.scene().addItem(self.viewboxes[EDILUTER]) # add viewbox to scene
-        self.viewboxes[EDILUTER].setXLink(self.plot) # link x axis of viewbox to x axis of plot
-        # eDiluter axis
-        self.axes[EDILUTER] = AxisItem('right') # create axis
-        self.plot.layout.addItem(self.axes[EDILUTER], 2, 8) # add axis to plot
-        self.axes[EDILUTER].setLabel('eDiluter temperature', units='°C', color='w') # set label
-        self.set_axis_style(self.axes[EDILUTER], 'w') # set axis style
-        self.axes[EDILUTER].linkToView(self.viewboxes[EDILUTER]) # link axis to viewbox
-
-        # Example device viewbox
-        self.viewboxes[EXAMPLE_DEVICE] = ViewBox() # create viewbox
-        self.plot.scene().addItem(self.viewboxes[EXAMPLE_DEVICE]) # add viewbox to scene
-        self.viewboxes[EXAMPLE_DEVICE].setXLink(self.plot) # link x axis of viewbox to x axis of plot
-        # Example device axis
-        self.axes[EXAMPLE_DEVICE] = AxisItem('right') # create axis
-        self.plot.layout.addItem(self.axes[EXAMPLE_DEVICE], 2, 9) # add axis to plot
-        self.axes[EXAMPLE_DEVICE].setLabel('Example device', units='units', color='w') # set label
-        self.set_axis_style(self.axes[EXAMPLE_DEVICE], 'w') # set axis style
-        self.axes[EXAMPLE_DEVICE].linkToView(self.viewboxes[EXAMPLE_DEVICE]) # link axis to viewbox
-        
-        # connect viewbox resize event to updateViews function
-        self.plot.vb.sigResized.connect(self.updateViews)
-        # call updateViews function to set viewboxes to same size
-        self.updateViews()
-
-        # connect plot's auto range button to set_auto_range function
-        self.plot.autoBtn.clicked.connect(self.set_auto_range)
-
-        # hide axes and disable SI scaling by default
-        for key in self.axes:
-            self.axes[key].hide()
-            self.axes[key].enableAutoSIPrefix(enable=False) # disable auto SI prefix
-        
-        # use automatic downsampling and clipping to reduce the drawing load
-        self.plot.setDownsampling(mode='peak')
-        self.plot.setClipToView(True)
-        # TODO does this affect all viewboxes?
-    
-    # handle view resizing
-    # called when plot widget (or window) is resized
-    # source: https://stackoverflow.com/questions/42931474/how-can-i-have-multiple-left-axisitems-with-the-same-alignment-position-using-py
-    def updateViews(self):
-        # set viewbox geometry to plot geometry
-        for viewbox in self.viewboxes.values():
-            # exclude CPC viewbox
-            if viewbox != self.viewboxes[CPC]:
-                viewbox.setGeometry(self.plot.vb.sceneBoundingRect())
-                # update linked axes
-                viewbox.linkedViewChanged(self.plot.vb, viewbox.XAxis)
-    
-    # set auto range on for all viewboxes
-    def set_auto_range(self):
-        for viewbox in self.viewboxes.values():
-            viewbox.enableAutoRange()
-    
-    def set_axis_style(self, axis, color):
-        axis.setStyle(tickFont=QFont("Arial", 12, QFont.Normal), tickLength=-20)
-        axis.setPen(color)
-        axis.setTextPen(color)
-        axis.label.setFont(QFont("Arial", 12, QFont.Normal)) # change axis label font
-
-    def show_hide_axis(self, device_type, show):
-        if device_type == PSM2:
-            axis = self.axes[PSM]
-        elif device_type == TSI_CPC:
-            axis = self.axes[CPC]
-        else:
-            axis = self.axes[device_type]
-        if show:
-            axis.show() # show axis
-        else:
-            axis.hide() # hide axis
-    
-    # change rhtp axis label according to value type
-    # None, "RH", "T", "P"
-    def change_rhtp_axis(self, value):
-        # TODO: only change axis if value differs from current axis
-        if value == None:
-            self.axes[RHTP].setLabel('RHTP', units=None, color='w')
-            self.axes[RHTP].hide() # hide axis
-        else:
-            if value == "RH":
-                self.axes[RHTP].setLabel('RHTP RH', units='%', color='w')
-            elif value == "T":
-                self.axes[RHTP].setLabel('RHTP T', units='°C', color='w')
-            elif value == "P":
-                self.axes[RHTP].setLabel('RHTP P', units='Pa', color='w')
-            self.axes[RHTP].show() # show axis
-        # set axis style
-        self.set_axis_style(self.axes[RHTP], 'w')
-    
-    def change_afm_axis(self, value):
-        if value == None:
-            self.axes[AFM].setLabel('AFM', units=None, color='w')
-            self.axes[AFM].hide() # hide axis
-        else:
-            if value == "Flow":
-                self.axes[AFM].setLabel('AFM flow', units='lpm', color='w')
-            elif value == "Standard flow":
-                self.axes[AFM].setLabel('AFM standard flow', units='slpm', color='w')
-            elif value == "RH":
-                self.axes[AFM].setLabel('AFM RH', units='%', color='w')
-            elif value == "T":
-                self.axes[AFM].setLabel('AFM T', units='°C', color='w')
-            elif value == "P":
-                self.axes[AFM].setLabel('AFM P', units='Pa', color='w')
-            self.axes[AFM].show() # show axis
-        # set axis style
-        self.set_axis_style(self.axes[AFM], 'w')
-        
-# triple plot widget containing three plots
-class TriplePlot(GraphicsLayoutWidget):
-    def __init__(self, device_type, *args, **kwargs):
-        super().__init__()
-
-        if device_type == 5: # RHTP
-            value_names = ["RH", "T", "P"]
-            unit_names = ["%", "°C", "Pa"]
-            colors = [(100,188,255), (255,100,100), (255, 255, 100)]
-        else: # other devices
-            value_names = ["val1", "val2", "val3"]
-            unit_names = ["unit1", "unit2", "unit3"]
-            colors = [(255,255,255), (255,100,100), (100,188,255)]
-        
-        # create plot by adding it to widget
-        self.plot = self.addPlot()
-
-        # time axis
-        self.plot.setAxisItems({'bottom':DateAxisItem()}) # set time axis to bottom
-        self.plot.setLabel('bottom', "Time") # set time axis label
-        self.axis_time = self.plot.getAxis('bottom') # store time axis to variable
-        self.set_axis_style(self.axis_time, 'w') # set axis style
-
-        # viewbox 1
-        self.viewbox1 = self.plot.getViewBox() # store default viewbox to variable
-        # axis 1
-        self.axis1 = self.plot.getAxis('left') # store left axis to variable
-        self.axis1.setLabel(value_names[0], units=unit_names[0], color=colors[0]) # set label
-        self.set_axis_style(self.axis1, colors[0]) # set axis style
-        # curve 1
-        self.curve1 = PlotCurveItem(pen=colors[0], connect="finite") # create curve 1
-        self.viewbox1.addItem(self.curve1) # add curve 1 to viewbox 1
-
-        # viewbox 2
-        self.viewbox2 = ViewBox() # create viewbox
-        self.plot.scene().addItem(self.viewbox2) # add viewbox to scene
-        self.viewbox2.setXLink(self.plot) # link x axis of viewbox to x axis of plot
-        # axis 2
-        self.axis2 = AxisItem('right') # create second axis
-        self.plot.layout.addItem(self.axis2, 2, 3) # add axis to plot
-        self.axis2.setLabel(value_names[1], units=unit_names[1], color=colors[1]) # set label
-        self.set_axis_style(self.axis2, colors[1]) # set axis style
-        self.axis2.linkToView(self.viewbox2) # link axis to viewbox
-        # curve 2
-        self.curve2 = PlotCurveItem(pen=colors[1], connect="finite") # create curve 2
-        self.viewbox2.addItem(self.curve2) # add curve 2 to viewbox 2
-
-        # viewbox 3
-        self.viewbox3 = ViewBox() # create viewbox
-        self.plot.scene().addItem(self.viewbox3) # add viewbox to scene
-        self.viewbox3.setXLink(self.plot) # link x axis of viewbox to x axis of plot
-        # axis 3
-        self.axis3 = AxisItem('right') # create third axis
-        self.plot.layout.addItem(self.axis3, 2, 4) # add axis to plot
-        self.axis3.setLabel(value_names[2], units=unit_names[2], color=colors[2]) # set label
-        self.set_axis_style(self.axis3, colors[2]) # set axis style
-        self.axis3.linkToView(self.viewbox3) # link axis to viewbox
-        # curve 3
-        self.curve3 = PlotCurveItem(pen=colors[2], connect="finite") # create curve 3
-        self.viewbox3.addItem(self.curve3) # add curve 3 to viewbox 3
-
-        # create list of viewboxes
-        self.viewboxes = [self.viewbox1, self.viewbox2, self.viewbox3]
-
-        # connect viewbox resize event to updateViews function
-        self.plot.vb.sigResized.connect(self.updateViews)
-        # call updateViews function to set viewboxes to same size
-        self.updateViews()
-        
-        # use automatic downsampling and clipping to reduce the drawing load
-        self.plot.setDownsampling(mode='peak')
-        self.plot.setClipToView(True)
-    
-    # handle view resizing
-    # called when plot widget (or window) is resized
-    # source: https://stackoverflow.com/questions/42931474/how-can-i-have-multiple-left-axisitems-with-the-same-alignment-position-using-py
-    def updateViews(self):
-        # set viewbox geometry to plot geometry
-        for viewbox in self.viewboxes[1:]: # exclude viewbox 1
-            viewbox.setGeometry(self.plot.vb.sceneBoundingRect())
-            # update linked axes
-            viewbox.linkedViewChanged(self.plot.vb, viewbox.XAxis)
-
-    def set_axis_style(self, axis, color):
-        axis.setStyle(tickFont=QFont("Arial", 12, QFont.Normal), tickLength=-20)
-        axis.setPen(color)
-        axis.setTextPen(color)
-        axis.label.setFont(QFont("Arial", 12, QFont.Normal)) # change axis label font
-        axis.enableAutoSIPrefix(enable=False) # disable auto SI prefix
-
-class AFMPlot(GraphicsLayoutWidget):
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-
-        # define value names, units and colors
-        value_names = ["Flow", "Standard flow", "RH", "T", "P"]
-        unit_names = ["lpm", "slpm", "%", "°C", "Pa"]
-        colors = [(255,255,255), (255,100,255), (100,188,255), (255, 100, 100), (255, 255, 100)]
-
-        # create plot item and hide its default axes
-        self.plot = PlotItem()
-        self.plot.hideAxis('left')
-        self.plot.hideAxis('bottom')
-        # add plot to widget column 2, leave space for left axes
-        self.addItem(self.plot, row=0, col=2)
-
-        # create viewboxes for each axis
-        self.viewboxes = []
-        # store default viewbox
-        self.viewboxes.append(self.plot.getViewBox())
-        # create 4 additional viewboxes
-        for i in range(4):
-            viewbox = ViewBox() # create viewbox
-            self.plot.scene().addItem(viewbox) # add viewbox to scene
-            viewbox.setXLink(self.plot) # link x axis of viewbox to x axis of plot
-            self.viewboxes.append(viewbox) # store viewbox to list
-        
-        # create axes for each viewbox
-        self.axes = []
-        for i in range(2):
-            axis = AxisItem('left') # create left axis
-            self.axes.append(axis) # store axis to list
-            self.addItem(axis, row=0, col=i) # add axis to widget
-        for i in range(3):
-            axis = AxisItem('right') # create right axis
-            self.axes.append(axis) # store axis to list
-            self.addItem(axis, row=0, col=i+3) # add axis to widget, skip left axes and plot
-        
-        # link axes to viewboxes
-        for i in range(5):
-            self.axes[i].linkToView(self.viewboxes[i])
-        
-        # set axis labels and styles
-        for i in range(5):
-            # set label
-            self.axes[i].setLabel(value_names[i], units=unit_names[i], color=colors[i])
-            # set style
-            self.axes[i].setStyle(tickFont=QFont("Arial", 12, QFont.Normal), tickLength=-20)
-            self.axes[i].setPen(colors[i])
-            self.axes[i].setTextPen(colors[i])
-            self.axes[i].label.setFont(QFont("Arial", 12, QFont.Normal)) # change axis label font
-            self.axes[i].enableAutoSIPrefix(enable=False) # disable auto SI prefix
-        
-        # create bottom time axis
-        self.axis_time = DateAxisItem('bottom')
-        # link time axis to viewbox
-        self.axis_time.linkToView(self.viewboxes[0])
-        # set botton axis label and style
-        self.axis_time.setLabel("Time")
-        self.axis_time.setStyle(tickFont=QFont("Arial", 12, QFont.Normal), tickLength=-20)
-        self.axis_time.setPen('w')
-        self.axis_time.setTextPen('w')
-        self.axis_time.label.setFont(QFont("Arial", 12, QFont.Normal)) # change axis label font
-        self.axis_time.enableAutoSIPrefix(enable=False) # disable auto SI prefix
-
-        # add bottom time axis to widget, same column as plot but on next row
-        self.addItem(self.axis_time, row=1, col=2)
-
-        # create curves for each viewbox
-        self.curves = []
-        for i in range(5):
-            curve = PlotCurveItem(pen=colors[i], connect="finite") # create curve
-            self.curves.append(curve) # store curve to list
-            self.viewboxes[i].addItem(curve) # add curve to viewbox
-        
-        # use automatic downsampling and clipping to reduce the drawing load
-        self.plot.setDownsampling(mode='peak')
-        self.plot.setClipToView(True)
-        
-        # connect viewbox resize event to updateViews function
-        self.plot.vb.sigResized.connect(self.updateViews)
-        # call updateViews function to set viewboxes to same size
-        self.updateViews()
-    
-    def updateViews(self):
-        # set viewbox geometry to plot geometry
-        for viewbox in self.viewboxes[1:]: # exclude first viewbox
-            viewbox.setGeometry(self.plot.vb.sceneBoundingRect())
-            # update linked axes
-            viewbox.linkedViewChanged(self.plot.vb, viewbox.XAxis)
-
-class ElectrometerPlot(GraphicsLayoutWidget):
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-
-        # list for storing references to plots
-        self.plots = []
-        # create plots and curves
-        # Voltage 1
-        self.plot1 = self.addPlot()
-        self.curve1 = self.plot1.plot(pen="g", connect="finite")
-        self.plots.append(self.plot1)
-        self.nextRow()
-        # Voltage 2
-        self.plot2 = self.addPlot()
-        self.curve2 = self.plot2.plot(pen="r", connect="finite")
-        self.plots.append(self.plot2)
-        self.nextRow()
-        # Voltage 3
-        self.plot3 = self.addPlot()
-        self.curve3 = self.plot3.plot(pen="b", connect="finite")
-        self.plots.append(self.plot3)
-        # set up plots
-        for i in range(3):
-            label = "Voltage " + str(i+1)
-            self.plots[i].setLabel('left', label, units='V') # set y-axis label
-            self.plots[i].setLabel('bottom', "Time") # set time axis label
-            self.plots[i].setAxisItems({'bottom':DateAxisItem()})
-            #self.plots[i].getAxis('left').enableAutoSIPrefix(enable=False) # disable auto SI prefix
-            self.plots[i].getAxis('bottom').enableAutoSIPrefix(enable=False) # disable auto SI prefix
-            self.plots[i].showGrid(x=True, y=True) # show grid by default
-            self.plots[i].setDownsampling(mode='peak')
-            self.plots[i].setClipToView(True)
-
-# single plot widget containing plot
-class SinglePlot(GraphicsLayoutWidget):
-    def __init__(self, device_type, *args, **kwargs):
-        super().__init__()
-
-        # create plot and curve
-        self.plot = self.addPlot() # create plot by adding it to widget
-        self.curve = self.plot.plot(pen="w", connect="finite") # create plot curve
-
-        # plot settings
-        self.plot.setAxisItems({'bottom':DateAxisItem()}) # set time axis to bottom
-        self.plot.getAxis('bottom').enableAutoSIPrefix(enable=False) # disable auto SI prefix 
-        self.plot.setLabel('bottom', "Time") # set time axis label
-        self.plot.showGrid(x=True, y=True) # show grid by default
-        self.plot.getAxis('left').enableAutoSIPrefix(enable=False) # disable auto SI prefix
-        # use automatic downsampling and clipping to reduce the drawing load
-        self.plot.setDownsampling(mode='peak')
-        self.plot.setClipToView(True)
-
-        # set y-axis label and units based on device type
-        if device_type == CPC:
-            self.plot.setLabel('left', "Concentration", units='#/cc')
-        elif device_type == PSM:
-            self.plot.setLabel('left', "Saturator flow", units='lpm')
-        elif device_type == CO2_SENSOR:
-            self.plot.setLabel('left', "CO2", units='ppm')
-        elif device_type == EDILUTER:
-            self.plot.setLabel('left', "eDiluter temperature", units='°C')
-        elif device_type == AFM:
-            self.plot.setLabel('left', "Flow", units='lpm')
-        elif device_type == EXAMPLE_DEVICE:
-            self.plot.setLabel('left', "Example device", units='units')
-        
-        self.viewbox = self.plot.getViewBox() # store viewbox to variable
-
-# CPC widget containing CPC related GUI elements as tabs
-class CPCWidget(QTabWidget):
-    def __init__(self, device_parameter, *args, **kwargs):
-        super().__init__()
-        self.device_parameter = device_parameter # store device parameter tree reference
-        self.name = device_parameter.name() # store device name
-        # create set tab widget for cpc settings
-        self.set_tab = CPCSetTab()
-        self.addTab(self.set_tab, "Set")
-        # create status tab widget showing CPC values
-        self.status_tab = CPCStatusTab()
-        self.addTab(self.status_tab, "Status")
-        # create plot widget for Concentration
-        self.plot_tab = SinglePlot(device_type=CPC)
-        self.addTab(self.plot_tab, "Concentration")
-        # create pulse quality widget for CPC pulse quality monitoring
-        self.pulse_quality = PulseQuality()
-        self.addTab(self.pulse_quality, "Pulse quality")
-
-        # create list of widget references for updating gui with cpc system status
-        self.cpc_status_widgets = [
-            self.status_tab.temp_optics, self.status_tab.temp_saturator,
-            self.status_tab.temp_condenser, self.status_tab.pres_inlet,
-            self.status_tab.pres_nozzle, self.status_tab.laser_power,
-            self.status_tab.liquid_level, self.status_tab.temp_cabin,
-            self.status_tab.pres_critical_orifice, self.status_tab.pulse_quality
-        ]
-
-    # convert CPC status hex to binary and update error label colors
-    def update_errors(self, status_hex, cabin_p_error):
-        widget_amount = len(self.cpc_status_widgets) # get amount of widgets
-        status_bin = bin(int(status_hex, 16)) # convert hex to int and int to binary
-        status_bin = status_bin[2:].zfill(widget_amount) # remove 0b from string and fill with 0s
-        total_errors = status_bin.count("1") # count number of 1s in status_bin
-        inverted_status_bin = status_bin[::-1] # invert status_bin for error parsing
-        for i in range(widget_amount): # iterate through all status widgets
-            # change color of error label according to error bit
-            self.cpc_status_widgets[i].change_color(inverted_status_bin[i])
-        # update cabin pressure label color according to error status
-        if cabin_p_error:
-            self.status_tab.pres_cabin.change_color(1)
-            total_errors += 1
-        else:
-            self.status_tab.pres_cabin.change_color(0)
-        
-        return total_errors # return total number of errors
-    
-    def update_settings(self, settings):
-        # update GUI set values if they differ from CPC set values
-        # TODO remove repetition
-
-        # saturator temperature
-        if self.set_tab.set_saturator_temp.value_spinbox.value() != settings[8]:
-            # update value
-            self.set_tab.set_saturator_temp.value_spinbox.setValue(settings[8])
-            # if saturator temperature is nan, clear visible value
-            if str(settings[8]) == 'nan':
-                self.set_tab.set_saturator_temp.value_spinbox.clear()
-            # if text is empty (without suffix), set text with value
-            # TODO ? change this to text().split(" ")[0] == ""
-            elif self.set_tab.set_saturator_temp.value_spinbox.text()[:-3] == "":
-                self.set_tab.set_saturator_temp.value_spinbox.lineEdit().setText(str(settings[8]))
-
-        # condenser temperature
-        if self.set_tab.set_condenser_temp.value_spinbox.value() != settings[6]:
-            # update value
-            self.set_tab.set_condenser_temp.value_spinbox.setValue(settings[6])
-            # if condenser temperature is nan, clear visible value
-            if str(settings[6]) == 'nan':
-                self.set_tab.set_condenser_temp.value_spinbox.clear()
-            # if text is empty (without suffix), set text with value
-            elif self.set_tab.set_condenser_temp.value_spinbox.text()[:-3] == "":
-                self.set_tab.set_condenser_temp.value_spinbox.lineEdit().setText(str(settings[6]))
-
-        # averaging time
-        if self.set_tab.set_averaging_time.value_spinbox.value() != settings[5]:
-            # update value
-            if str(settings[5]) == 'nan': # if nan, set to 0
-                self.set_tab.set_averaging_time.value_spinbox.setValue(0)
-            else: # else update value
-                self.set_tab.set_averaging_time.value_spinbox.setValue(settings[5])
-            # if averaging time is nan, clear visible value
-            if str(settings[5]) == 'nan':
-                self.set_tab.set_averaging_time.value_spinbox.clear()
-            # if text is empty (without suffix), update value set text with value
-            elif self.set_tab.set_averaging_time.value_spinbox.text()[:-2] == "":
-                self.set_tab.set_averaging_time.value_spinbox.lineEdit().setText(str(settings[5]))
-        
-        # update mode settings
-        self.set_tab.autofill.update_state(settings[1]) # autofill
-        self.set_tab.water_removal.update_state(settings[4]) # water removal
-        self.set_tab.drain.update_state(settings[2]) # drain
-    
-    # update all data values in status tab
-    def update_values(self, current_list):
-        # update temperature values
-        self.status_tab.temp_optics.change_value(str(current_list[5]) + " °C")
-        self.status_tab.temp_saturator.change_value(str(current_list[3]) + " °C")
-        self.status_tab.temp_condenser.change_value(str(current_list[4]) + " °C")
-        # update pressure values
-        self.status_tab.pres_inlet.change_value(str(current_list[7]) + " kPa")
-        self.status_tab.pres_nozzle.change_value(str(current_list[9]) + " kPa")
-        self.status_tab.pres_critical_orifice.change_value(str(current_list[8]) + " kPa")
-        self.status_tab.pres_cabin.change_value(str(current_list[10]) + " kPa")
-        # update misc values
-        if current_list[11] == 0:
-            self.status_tab.liquid_level.change_value("LOW")
-        elif current_list[11] == 1:
-            self.status_tab.liquid_level.change_value("OK")
-        elif current_list[11] == 2:
-            self.status_tab.liquid_level.change_value("OVERFILL")
-        self.status_tab.temp_cabin.change_value(str(current_list[6]) + " °C")
-
-# PSM widget
-class PSMWidget(QTabWidget):
-    def __init__(self, device_parameter, device_type, *args, **kwargs):
-        super().__init__()
-        self.device_parameter = device_parameter # store device parameter tree reference
-        self.name = device_parameter.name() # store device name
-        self.device_type = device_type # store device type (PSM or PSM 2.0)
-        # create set tab for PSM
-        self.set_tab = PSMSetTab(device_type)
-        self.addTab(self.set_tab, "Set")
-        # create status tab for PSM
-        self.status_tab = PSMStatusTab(device_type)
-        self.addTab(self.status_tab, "Status")
-        # create mode tab for PSM
-        self.measure_tab = PSMMeasureTab()
-        self.addTab(self.measure_tab, "Measure")
-        # create plot widget for PSM
-        self.plot_tab = SinglePlot(device_type=PSM)
-        self.addTab(self.plot_tab, "PSM plot")
-
-        # create list of PSM status widgets, used in update_errors
-        self.psm_status_widgets = [
-            self.status_tab.temp_growth_tube, self.status_tab.temp_saturator,
-            self.status_tab.flow_saturator, self.status_tab.temp_heater,
-            self.status_tab.temp_inlet, "mix1_press", "mix2_press",
-            self.status_tab.pressure_inlet, self.status_tab.flow_excess,
-            "drain_level", self.status_tab.temp_cabin, self.status_tab.temp_drainage,
-            self.status_tab.pressure_critical_orifice, "mfc_temp"
-        ]
-        # if PSM 2.0, add vacuum flow widget to list
-        if device_type == PSM2:
-            self.psm_status_widgets.append(self.status_tab.flow_vacuum)
-
-    # convert PSM status hex to binary and update error label colors
-    def update_errors(self, status_hex):
-        widget_amount = len(self.psm_status_widgets) # get amount of widgets in list
-        status_bin = bin(int(status_hex, 16)) # convert hex to int and int to binary
-        status_bin = status_bin[2:].zfill(widget_amount) # remove 0b from string and fill with 0s to length of widget_amount
-        total_errors = status_bin.count("1") # count number of 1s in status_bin
-        inverted_status_bin = status_bin[::-1] # invert status_bin for error parsing
-        for i in range(widget_amount): # iterate through all status widgets
-            if type(self.psm_status_widgets[i]) != str: # filter placeholder strings
-                # change color of error label according to error bit
-                self.psm_status_widgets[i].change_color(inverted_status_bin[i])
-        
-        return total_errors # return total number of errors
-    
-    # convert PSM notes hex to binary and update liquid mode settings
-    def update_notes(self, note_hex):
-        liquid_errors = 0 # increment if liquid errors occur
-        note_length = 7 # if new note bits are added in firmware, change this value accordingly
-        note_bin = bin(int(note_hex, 16)) # convert hex to int and int to binary
-        note_bin = note_bin[2:].zfill(note_length) # remove 0b from string and fill with 0s
-        total_notes = note_bin.count("1") # count number of 1s in note_bin
-        inverted_note_bin = note_bin[::-1] # invert note_bin for liquid setting parsing
-        # update liquid mode settings in GUI
-        # 0 = autofill on, 1 = autofill off
-        if inverted_note_bin[5] == "0":
-            self.set_tab.autofill.update_state(1)
-        elif inverted_note_bin[5] == "1":
-            self.set_tab.autofill.update_state(0)
-        # 0 = drying off, 1 = drying on
-        self.set_tab.drying.update_state(int(inverted_note_bin[4]))
-        # 0 = drain on, 1 = drain off
-        if inverted_note_bin[3] == "0":
-            self.set_tab.drain.update_state(1)
-        elif inverted_note_bin[3] == "1":
-            self.set_tab.drain.update_state(0)
-        # 0 = saturator liquid level OK, 1 = saturator liquid level LOW
-        self.status_tab.liquid_saturator.change_color(inverted_note_bin[6])
-        if inverted_note_bin[6] == "1":
-            liquid_errors += 1
-        # 0 = drain liquid level OK, 1 = drain liquid level HIGH
-        self.status_tab.liquid_drain.change_color(inverted_note_bin[0])
-        if inverted_note_bin[0] == "1":
-            liquid_errors += 1
-
-        return liquid_errors # return total number of liquid errors
-
-    def update_settings(self, settings):
-        self.set_tab.set_growth_tube_temp.value_spinbox.setValue(float(settings[1]))
-        self.set_tab.set_saturator_temp.value_spinbox.setValue(float(settings[2]))
-        self.set_tab.set_inlet_temp.value_spinbox.setValue(float(settings[3]))
-        self.set_tab.set_heater_temp.value_spinbox.setValue(float(settings[4]))
-        self.set_tab.set_drainage_temp.value_spinbox.setValue(float(settings[5]))
-        self.set_tab.set_cpc_inlet_flow.value_spinbox.setValue(float(settings[6]))
-    
-    # update all data values in status tab
-    def update_values(self, current_list):
-        # update temperature values
-        self.status_tab.temp_growth_tube.change_value(str(current_list[2]) + " °C")
-        self.status_tab.temp_saturator.change_value(str(current_list[3]) + " °C")
-        self.status_tab.temp_inlet.change_value(str(current_list[4]) + " °C")
-        self.status_tab.temp_heater.change_value(str(current_list[5]) + " °C")
-        self.status_tab.temp_drainage.change_value(str(current_list[6]) + " °C")
-        self.status_tab.temp_cabin.change_value(str(current_list[7]) + " °C")
-        # update flow values
-        # self.status.flow_cpc is updated in PSMWidget's update_settings()
-        self.status_tab.flow_saturator.change_value(str(current_list[0]) + " lpm")
-        self.status_tab.flow_excess.change_value(str(current_list[1]) + " lpm")
-        # self.status_tab.flow_inlet is updated in update_plot_data()
-        # update pressure values
-        self.status_tab.pressure_inlet.change_value(str(current_list[9]) + " kPa")
-        self.status_tab.pressure_critical_orifice.change_value(str(current_list[12]) + " kPa")
-        # update vacuum flow if PSM 2.0
-        if self.device_type == PSM2:
-            self.status_tab.flow_vacuum.change_value(str(current_list[13]) + " lpm")
-        # liquid level values are updated in PSMWidget's update_notes()
-
-class PSMSetTab(QSplitter):
-    def __init__(self, device_type, *args, **kwargs):
-        super().__init__()
-        # split tab vertically
-        self.setOrientation(Qt.Vertical)
-
-        # TODO check device type and create widgets accordingly
-
-        # horizontal splitter containing upper half of tab - set widgets
-        upper_splitter = QSplitter(Qt.Horizontal)
-        self.set_growth_tube_temp = SetWidget("Growth tube T", " °C")
-        upper_splitter.addWidget(self.set_growth_tube_temp)
-        self.set_saturator_temp = SetWidget("Saturator T", " °C")
-        upper_splitter.addWidget(self.set_saturator_temp)
-        self.set_inlet_temp = SetWidget("Inlet T", " °C")
-        upper_splitter.addWidget(self.set_inlet_temp)
-        self.set_heater_temp = SetWidget("Heater T", " °C")
-        upper_splitter.addWidget(self.set_heater_temp)
-        self.set_drainage_temp = SetWidget("Drainage T", " °C")
-        upper_splitter.addWidget(self.set_drainage_temp)
-        # horizontal splitter containing middle half of tab - set widgets
-        middle_splitter = QSplitter(Qt.Horizontal)
-        self.set_cpc_inlet_flow = SetWidget("CPC inlet flow rate\n(used in dilution correction)", " lpm", decimals=3)
-        middle_splitter.addWidget(self.set_cpc_inlet_flow)
-        self.set_cpc_sample_flow = SetWidget("CPC sample flow rate\n(used in concentration calculation)", " lpm", decimals=3)
-        middle_splitter.addWidget(self.set_cpc_sample_flow)
-        if device_type == PSM: # if PSM, add CO flow rate set widget
-            self.set_co_flow = SetWidget("CO flow rate", " lpm", decimals=3)
-            middle_splitter.addWidget(self.set_co_flow)
-        # horizontal splitter containing lower half of tab - mode widgets
-        lower_splitter = QSplitter(Qt.Horizontal)
-        self.autofill = ToggleButton("Autofill")
-        lower_splitter.addWidget(self.autofill)
-        self.drain = ToggleButton("Drain")
-        lower_splitter.addWidget(self.drain)
-        self.drying = ToggleButton("Drying")
-        lower_splitter.addWidget(self.drying)
-        # set splitter's relative widget sizes and add to tab
-        upper_splitter.setSizes([1000, 1000, 1000, 1000, 1000])
-        self.addWidget(upper_splitter)
-        middle_splitter.setSizes([1000, 1000, 1000])
-        self.addWidget(middle_splitter)
-        lower_splitter.setSizes([1000, 1000, 1000])
-        self.addWidget(lower_splitter)
-        # add line edit for command input
-        if device_type == PSM: # if PSM
-            self.command_widget = CommandWidget("PSM Retrofit")
-        elif device_type == PSM2: # if PSM 2.0
-            self.command_widget = CommandWidget("PSM 2.0")
-        self.addWidget(self.command_widget)
-        # set relative sizes in tab splitter
-        self.setSizes([1000, 1000, 1000, 1000])
-    
-class PSMStatusTab(QWidget):
-    def __init__(self, device_type, *args, **kwargs):
-        super().__init__()
-
-        layout = QGridLayout() # create layout
-
-        # TODO check device type and create widgets accordingly
-
-        # temperature indicators
-        self.temp_growth_tube = IndicatorWidget("Growth tube temperature")
-        layout.addWidget(self.temp_growth_tube, 0, 0)
-        self.temp_saturator = IndicatorWidget("Saturator temperature")
-        layout.addWidget(self.temp_saturator, 1, 0)
-        self.temp_inlet = IndicatorWidget("Inlet temperature")
-        layout.addWidget(self.temp_inlet, 2, 0)
-        self.temp_heater = IndicatorWidget("Heater temperature")
-        layout.addWidget(self.temp_heater, 3, 0)
-        self.temp_drainage = IndicatorWidget("Drainage temperature")
-        layout.addWidget(self.temp_drainage, 4, 0)
-        self.temp_cabin = IndicatorWidget("Cabin temperature")
-        layout.addWidget(self.temp_cabin, 0, 1)
-
-        # flow indicators
-        self.flow_cpc = IndicatorWidget("CPC inlet flow")
-        layout.addWidget(self.flow_cpc, 1, 1)
-        self.flow_saturator = IndicatorWidget("Saturator flow")
-        layout.addWidget(self.flow_saturator, 2, 1)
-        self.flow_excess = IndicatorWidget("Excess flow") # TODO change name to heater flow?
-        layout.addWidget(self.flow_excess, 3, 1)
-        self.flow_inlet = IndicatorWidget("Inlet flow")
-        layout.addWidget(self.flow_inlet, 4, 1)
-        if device_type == PSM2: # if PSM 2.0, add vacuum flow indicator
-            self.flow_vacuum = IndicatorWidget("Vacuum flow")
-            layout.addWidget(self.flow_vacuum, 4, 2)
-
-        # pressure indicators
-        self.pressure_inlet = IndicatorWidget("Inlet pressure")
-        layout.addWidget(self.pressure_inlet, 0, 2)
-        if device_type == PSM: # if PSM, add critical orifice pressure indicator
-            self.pressure_critical_orifice = IndicatorWidget("Critical orifice pressure")
-            layout.addWidget(self.pressure_critical_orifice, 1, 2)
-        elif device_type == PSM2: # if PSM 2.0, add vacuum line pressure indicator
-            # TODO name variable accordingly?
-            self.pressure_critical_orifice = IndicatorWidget("Vacuum line pressure")
-            layout.addWidget(self.pressure_critical_orifice, 1, 2)
-
-        # liquid level indicators
-        self.liquid_saturator = IndicatorWidget("Saturator liquid level")
-        layout.addWidget(self.liquid_saturator, 2, 2)
-        self.liquid_drain = IndicatorWidget("Drain liquid level")
-        layout.addWidget(self.liquid_drain, 3, 2)
-
-        self.setLayout(layout)
-
-class PSMMeasureTab(QWidget):
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-
-        layout = QGridLayout() # create layout
-
-        # scan mode widgets
-        self.scan = StartButton("Scan")
-        layout.addWidget(self.scan, 0, 0)
-        self.set_minimum_flow = SetWidget("Minimum flow", " lpm")
-        self.set_minimum_flow.value_spinbox.setValue(0.15)
-        layout.addWidget(self.set_minimum_flow, 1, 0)
-        self.set_max_flow = SetWidget("Maximum flow", " lpm")
-        self.set_max_flow.value_spinbox.setValue(1.9)
-        layout.addWidget(self.set_max_flow, 2, 0)
-        self.set_scan_time = SetWidget("Scan time", " s", integer=True)
-        self.set_scan_time.value_spinbox.setValue(240)
-        layout.addWidget(self.set_scan_time, 3, 0)
-
-        # step mode widgets
-        self.step = StartButton("Step")
-        layout.addWidget(self.step, 0, 1)
-        self.step_time = SetWidget("Step time", " s", integer=True)
-        self.step_time.value_spinbox.setValue(30)
-        layout.addWidget(self.step_time, 1, 1)
-        self.steps = StepsWidget()
-        self.steps.text_box.setText("0.1\n0.7\n1.3\n1.9")
-        layout.addWidget(self.steps, 2, 1, 2, 1)
-
-        # fixed mode widgets
-        self.fixed = StartButton("Fixed")
-        layout.addWidget(self.fixed, 0, 2)
-        self.set_flow = SetWidget("Saturator flow", " lpm")
-        self.set_flow.value_spinbox.setValue(1.9)
-        layout.addWidget(self.set_flow, 1, 2)
-
-        # 10 hz logging button
-        self.ten_hz = StartButton("10 Hz logging")
-        layout.addWidget(self.ten_hz, 4, 0, 1, 3)
-        # set button size policy to minimum
-        self.ten_hz.setSizePolicy(QSizePolicy(QSizePolicy.Minimum, QSizePolicy.Minimum))
-
-        self.setLayout(layout)
-    
-    def compile_scan(self): # compile scan command
-        scan_time = self.set_scan_time.value_spinbox.value()
-        if scan_time % 2 == 0: # if scan time is even
-            time = int((scan_time - 20) / 2)
-            parameters = [10, time, 10, time]
-        else: # if scan time is odd
-            time = int((scan_time - 21) / 2)
-            parameters = [11, time, 10, time]
-        # add minimum flow to parameters
-        parameters.append(round(self.set_minimum_flow.value_spinbox.value(), 3))
-        # add maximum flow to parameters
-        parameters.append(round(self.set_max_flow.value_spinbox.value(), 3))
-        scan_string = ":SET:FLOW:SCAN " + ",".join(map(str, parameters))
-        print(scan_string)
-        return scan_string
-    
-    def compile_step(self): # compile step command
-        step_list = self.steps.text_box.toPlainText().split("\n") # get list of steps
-        while "" in step_list:
-            step_list.remove("") # remove empty rows
-        error_flag = False
-        self.steps.text_box.clear()
-        self.steps.text_box.setTextColor(self.steps.default_color)
-        for step in step_list: # remove non-float values from list
-            try:
-                float(step) # check if float
-                self.steps.text_box.append(step)
-            except ValueError:
-                # write rows containing errors with red text
-                self.steps.text_box.setTextColor(QColor(255, 0, 0))
-                self.steps.text_box.append(step)
-                self.steps.text_box.setTextColor(self.steps.default_color)
-                error_flag = True # set error flag
-        if error_flag: # if there are errors
-            return None
-        else:
-            step_amount = len(step_list)
-            step_times = [self.step_time.value_spinbox.value()] * step_amount
-            step_string = ":SET:FLOW:STEP " + str(step_amount) + "," + ",".join(map(str, step_times)) + "," + ",".join(map(str, step_list))
-            return step_string
-
-    def compile_fixed(self): # compile fixed command
-        # append saturator flow value to command
-        fixed_string = ":SET:FLOW:FXD " + str(round(self.set_flow.value_spinbox.value(), 3))
-        return fixed_string
-    
-    # change color of active mode
-    def change_mode_color(self, command):
-        # TODO only update if command is different from current
-        if command == ":MEAS:SCAN":
-            self.scan.change_color(1)
-            self.step.change_color(0)
-            self.fixed.change_color(0)
-        if command == ":MEAS:STEP":
-            self.scan.change_color(0)
-            self.step.change_color(1)
-            self.fixed.change_color(0)
-        if command == ":MEAS:FIXD":
-            self.scan.change_color(0)
-            self.step.change_color(0)
-            self.fixed.change_color(1)
-
-
-# ELECTROMETER widget
-class ElectrometerWidget(QTabWidget):
-    def __init__(self, device_parameter, *args, **kwargs):
-        super().__init__()
-        self.device_parameter = device_parameter # store device parameter tree reference
-        self.name = device_parameter.name() # store device name
-        # create plot widget for Electrometer
-        self.plot_tab = ElectrometerPlot()
-        self.addTab(self.plot_tab, "Electrometer plot")
-
-# CO2 widget
-class CO2Widget(QTabWidget):
-    def __init__(self, device_parameter, *args, **kwargs):
-        super().__init__()
-        self.device_parameter = device_parameter # store device parameter tree reference
-        self.name = device_parameter.name() # store device name
-        # create plot widget for CO2
-        self.plot_tab = SinglePlot(device_type=CO2_SENSOR)
-        self.addTab(self.plot_tab, "CO2 plot")
-
-# RHTP widget
-class RHTPWidget(QTabWidget):
-    def __init__(self, device_parameter, *args, **kwargs):
-        super().__init__()
-        self.device_parameter = device_parameter # store device parameter tree reference
-        self.name = device_parameter.name() # store device name
-        # create plot widget for RHTP
-        self.plot_tab = TriplePlot(device_type=RHTP)
-        self.addTab(self.plot_tab, "RHTP plot")
-
-# AFM widget
-class AFMWidget(QTabWidget):
-    def __init__(self, device_parameter, *args, **kwargs):
-        super().__init__()
-        self.device_parameter = device_parameter # store device parameter tree reference
-        self.name = device_parameter.name() # store device name
-        # create plot widget for AFM
-        self.plot_tab = AFMPlot()
-        self.addTab(self.plot_tab, "AFM plot")
-
-# eDiluter widget
-class eDiluterWidget(QTabWidget):
-    def __init__(self, device_parameter, *args, **kwargs):
-        super().__init__()
-        self.device_parameter = device_parameter # store device parameter tree reference
-        self.name = device_parameter.name() # store device name
-        self.current_mode = None # used for storing current mode
-        # create set tab for eDiluter
-        self.set_tab = eDiluterSetTab()
-        self.addTab(self.set_tab, "Set")
-        # create status tab for eDiluter
-        self.status_tab = eDiluterStatusTab()
-        self.addTab(self.status_tab, "Status")
-        # create plot widget for eDiluter
-        self.plot_tab = SinglePlot(device_type=EDILUTER)
-        self.addTab(self.plot_tab, "eDiluter plot")
-        # create dictionary with mode names and corresponding widgets
-        self.mode_dict = {"INIT": self.set_tab.init, "WARMUP": self.set_tab.warmup,
-                          "STANDBY": self.set_tab.standby, "MEASUREMENT": self.set_tab.measurement}
-    
-    # update all data values in status tab and set tab
-    # current list: Status, P1, P2, T1, T2, T3, T4, T5, T6, DF1, DF2, DFTot
-    def update_values(self, current_list):
-        # update temperature values
-        self.status_tab.t1.change_value(str(current_list[3]) + " °C")
-        self.status_tab.t2.change_value(str(current_list[4]) + " °C")
-        self.status_tab.t3.change_value(str(current_list[5]) + " °C")
-        self.status_tab.t4.change_value(str(current_list[6]) + " °C")
-        self.status_tab.t5.change_value(str(current_list[7]) + " °C")
-        self.status_tab.t6.change_value(str(current_list[8]) + " °C")
-        # update pressure values
-        self.status_tab.p1.change_value(str(current_list[1]) + " mbar")
-        self.status_tab.p2.change_value(str(current_list[2]) + " mbar")
-        # update dilution factor values
-        self.set_tab.df_1.indicator.change_value(str(current_list[9]))
-        self.set_tab.df_2.indicator.change_value(str(current_list[10]))
-        self.set_tab.df_tot.change_value(str(current_list[11])) # total DF in set tab
-        self.status_tab.df_tot.change_value(str(current_list[11])) # total DF in status tab
-        # change color of active mode if it differs from current mode
-        if current_list[0] != self.current_mode:
-            # change color of all mode buttons to default
-            for mode in self.mode_dict:
-                self.mode_dict[mode].change_color(0)
-            # if current mode is not nan
-            if str(current_list[0]) != "nan":
-                # change color of active mode button
-                self.mode_dict[current_list[0]].change_color(1)
-                self.current_mode = current_list[0] # update current mode
-
-# TSI CPC widget
-class TSIWidget(QTabWidget):
-    def __init__(self, device_parameter, *args, **kwargs):
-        super().__init__()
-        self.device_parameter = device_parameter # store device parameter tree reference
-        self.name = device_parameter.name() # store device name
-        # create plot widget for TSI CPC
-        self.plot_tab = SinglePlot(device_type=CPC)
-        self.addTab(self.plot_tab, "TSI CPC plot")
-
-# Example device widget
-class ExampleDeviceWidget(QTabWidget):
-    def __init__(self, device_parameter, *args, **kwargs):
-        super().__init__()
-        self.device_parameter = device_parameter # store device parameter tree reference
-        self.name = device_parameter.name() # store device name
-        # create plot widget for CO2
-        self.plot_tab = SinglePlot(device_type=EXAMPLE_DEVICE)
-        self.addTab(self.plot_tab, "Example device plot")
-
-class eDiluterSetTab(QWidget):
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-
-        layout = QVBoxLayout() # create vertical layout
-
-        # horizontal layout for mode settings
-        # INIT, WARMUP, STANDBY, MEASUREMENT
-        mode_layout = QHBoxLayout()
-        self.init = StartButton("Init")
-        mode_layout.addWidget(self.init)
-        self.warmup = StartButton("Warmup")
-        mode_layout.addWidget(self.warmup)
-        self.standby = StartButton("Standby")
-        mode_layout.addWidget(self.standby)
-        self.measurement = StartButton("Measurement")
-        mode_layout.addWidget(self.measurement)
-        layout.addLayout(mode_layout) # add mode_layout to main vertical layout
-
-        # horizontal layout for dilution factor settings
-        df_layout = QHBoxLayout()
-        self.df_1 = DFSetWidget("Dilution factor 1")
-        df_layout.addWidget(self.df_1)
-        self.df_2 = DFSetWidget("Dilution factor 2")
-        df_layout.addWidget(self.df_2)
-        self.df_tot = IndicatorWidget("Total dilution factor")
-        df_layout.addWidget(self.df_tot)
-        layout.addLayout(df_layout) # add df_layout to main vertical layout
-
-        # add line edit for command input
-        self.command_widget = CommandWidget("eDiluter")
-        layout.addWidget(self.command_widget)
-
-        self.setLayout(layout) # add layout to widget
-
-class DFSetWidget(QWidget):
-    def __init__(self, name, *args, **kwargs):
-        super().__init__()
-        layout = QHBoxLayout() # create horizontal layout
-
-        # create previous button
-        self.prev_button = QPushButton("<", objectName="button_widget")
-        # create indicator widget for displaying name and value
-        self.indicator = IndicatorWidget(name)
-        # create next button
-        self.next_button = QPushButton(">", objectName="button_widget")
-
-        # add widgets to layout
-        layout.addWidget(self.prev_button) # previous button
-        layout.addWidget(self.indicator) # indicator widget
-        layout.addWidget(self.next_button) # next button
-
-        self.setLayout(layout) # add layout to widget
-
-class eDiluterStatusTab(QWidget):
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-
-        layout = QGridLayout() # create layout
-
-        # temperature indicators (unit = °C)
-        self.t1 = IndicatorWidget("Dilution air temperature") # T1
-        layout.addWidget(self.t1, 0, 0)
-        self.t2 = IndicatorWidget("Ext 1 temperature") # T2
-        layout.addWidget(self.t2, 1, 0)
-        self.t3 = IndicatorWidget("Ext 2 temperature") # T3
-        layout.addWidget(self.t3, 2, 0)
-        self.t6 = IndicatorWidget("Internal temperature") # T6
-        layout.addWidget(self.t6, 0, 1)
-        self.t4 = IndicatorWidget("Aux 1 temperature") # T4
-        layout.addWidget(self.t4, 1, 1)
-        self.t5 = IndicatorWidget("Aux 2 temperature") # T5
-        layout.addWidget(self.t5, 2, 1)
-
-        # pressure indicators (unit = mbar)
-        self.p1 = IndicatorWidget("Inlet pressure") # P1
-        layout.addWidget(self.p1, 0, 2)
-        self.p2 = IndicatorWidget("Ambient pressure") # P2
-        layout.addWidget(self.p2, 1, 2)
-
-        # dilution factor indicator
-        self.df_tot = IndicatorWidget("Total dilution factor") # DFTot
-        layout.addWidget(self.df_tot, 2, 2)
-        
-        self.setLayout(layout) # set layout
-
-# set tab widget containing settings and message input
-# used in CPCWidget
-class CPCSetTab(QWidget):
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-
-        layout = QGridLayout() # create grid layout
-
-        self.set_saturator_temp = SetWidget("Saturator temperature", " °C")
-        layout.addWidget(self.set_saturator_temp, 0, 0)
-        self.set_condenser_temp = SetWidget("Condenser temperature", " °C")
-        layout.addWidget(self.set_condenser_temp, 0, 1)
-        self.set_averaging_time = SetWidget("Averaging time", " s")
-        layout.addWidget(self.set_averaging_time, 0, 2)
-        
-        self.autofill = ToggleButton("Autofill")
-        layout.addWidget(self.autofill, 1, 0)
-        self.water_removal = ToggleButton("Water removal")
-        layout.addWidget(self.water_removal, 1, 1)
-        self.drain = ToggleButton("Drain")
-        layout.addWidget(self.drain, 1, 2)
-        
-        # add line edit for command input
-        self.command_widget = CommandWidget("CPC")
-        layout.addWidget(self.command_widget, 2, 0, 1, 3)
-
-        layout.setRowStretch(0, 1) # set stretch factor of row 0
-        layout.setRowStretch(1, 1) # set stretch factor of row 1
-        layout.setRowStretch(2, 2) # set stretch factor of row 2
-
-        self.setLayout(layout) # add layout to widget
-
-
-
-# status tab containing status indicator widgets
-# used in CPCWidget
-class CPCStatusTab(QWidget):
-    def __init__(self, *args, **kwargs):
-        super().__init__()
-
-        layout = QGridLayout() # create layout
-
-        # temperature indicators
-        self.temp_optics = IndicatorWidget("Optics temperature") # create optics temperature indicator
-        layout.addWidget(self.temp_optics, 0, 0)
-        self.temp_saturator = IndicatorWidget("Saturator temperature") # create saturator temperature indicator
-        layout.addWidget(self.temp_saturator, 1, 0)
-        self.temp_condenser = IndicatorWidget("Condenser temperature") # create condenser temperature indicator
-        layout.addWidget(self.temp_condenser, 2, 0)
-        self.temp_cabin = IndicatorWidget("Cabin temperature") # create cabin temp indicator
-        layout.addWidget(self.temp_cabin, 3, 0)
-
-        # pressure indicators
-        self.pres_inlet = IndicatorWidget("Inlet pressure") # create inlet pressure indicator
-        layout.addWidget(self.pres_inlet, 0, 1)
-        self.pres_nozzle = IndicatorWidget("Nozzle pressure") # create nozzle pressure indicator
-        layout.addWidget(self.pres_nozzle, 1, 1)
-        self.pres_critical_orifice = IndicatorWidget("Critical orifice pressure") # create nozzle pressure indicator
-        layout.addWidget(self.pres_critical_orifice, 2, 1)
-        self.pres_cabin = IndicatorWidget("Cabin pressure") # create cabin pressure indicator
-        layout.addWidget(self.pres_cabin, 3, 1)
-
-        # misc indicators
-        self.laser_power = IndicatorWidget("Laser power") # create laser power indicator
-        layout.addWidget(self.laser_power, 0, 2, 1, 1)
-        self.liquid_level = IndicatorWidget("Liquid level") # create liquid level indicator
-        layout.addWidget(self.liquid_level, 1, 2, 1, 1)
-        self.pulse_quality = IndicatorWidget("Pulse quality") # create pulse quality indicator
-        layout.addWidget(self.pulse_quality, 2, 2, 1, 1)
-
-        self.setLayout(layout)
 
 
 # application format
