+++ src/devices/__init__.py
@@ -1,0 +1,9 @@
+from .cpc import CPCWidget
+from .psm import PSMWidget
+from .co2 import CO2Widget
+from .electrometer import ElectrometerWidget
+from .rhtp import RHTPWidget
+from .ediluter import eDiluterWidget
+from .afm import AFMWidget
+from .tsi_cpc import TSIWidget
+from .example import ExampleDeviceWidget
