+++ src/plots/__init__.py
@@ -1,0 +1,2 @@
+from .main_plot import MainPlot
+from .device_plots import TriplePlot, AFMPlot, ElectrometerPlot, SinglePlot
